refactoring updates for output_checkers
