+++ geoips/plugins/modules/output_checkers/geotiff.py
@@ -20,7 +20,7 @@ LOG = logging.getLogger(__name__)
 
 interface = "output_checkers"
 family = "standard"
-name = "geotiff_checker"
+name = "geotiff"
 
 
 def correct_type(fname):
@@ -41,11 +41,13 @@ def correct_type(fname):
     return False
 
 
-def outputs_match(output_product, compare_product):
+def outputs_match(plugin, output_product, compare_product):
     """Use diff system command to compare currently produced image to correct image.
 
     Parameters
     ----------
+    plugin: OutputCheckerPlugin
+        The correspdonding geotiff output_checker - not used but needed in signature
     output_product : str
         Full path to current output product
     compare_product : str
@@ -77,7 +79,7 @@ def outputs_match(output_product, compare_product):
     return True
 
 
-def call(plugin, compare_path, output_products, test_product_func=None):
+def call(plugin, compare_path, output_products):
     """Compare the "correct" geotiffs found the list of current output_products.
 
     Compares files produced in the current processing run with the list of
@@ -92,23 +94,11 @@ def call(plugin, compare_path, output_products, test_product_func=None):
     output_products : list of str
         List of strings of current output products,
         to compare with products in compare_path
-    test_product_func : function, default=None
-        Alternative function to be used for testing output product
-
-          * Call signature must be:
-
-              * output_product, compare_product, goodcomps, badcomps, compare_strings
-
-          * Return must be:
-
-              * goodcomps, badcomps, compare_strings
-
-        * If None, use geoips.compare_outputs.test_product)
 
     Returns
     -------
     int
         Binary code: 0 if all comparisons were completed successfully.
     """
-    retval = plugin.compare_outputs(compare_path, output_products, test_product_func)
+    retval = plugin.compare_outputs(compare_path, output_products)
     return retval
